Break segmentFunctions into separate files
