+++ src/segmentFunctions/segmentGetLengthSq.ts
@@ -1,0 +1,7 @@
+import { ISegment } from "../types";
+
+export function segmentGetLengthSq(segment: ISegment) {
+  const dx = segment.x1 - segment.x0;
+  const dy = segment.y1 - segment.y0;
+  return dx * dx + dy * dy;
+}
