+++ src/segmentFunctions/segmentIntersectRay.ts
@@ -1,0 +1,8 @@
+import { _intersectionSwap } from "../internal/internalFunctions";
+import { intersectionAlloc } from "../intersectionFunctions/intersectionAlloc";
+import { rayIntersectSegment } from "../rayFunctions/rayIntersectSegment";
+import { IRay, ISegment } from "../types";
+
+export function segmentIntersectRay(segment: ISegment, ray: IRay, out = intersectionAlloc()) {
+  return _intersectionSwap(rayIntersectSegment(ray, segment, out));
+}
